Move EnvironmentLog to interface.
