+++ src/main/java/io/adobe/cloudmanager/Environment.java
@@ -23,7 +23,6 @@ package io.adobe.cloudmanager;
 import java.io.File;
 import java.util.List;
 
-import io.adobe.cloudmanager.model.EnvironmentLog;
 import io.adobe.cloudmanager.model.Variable;
 
 public interface Environment {
