+++ src/test/java/io/adobe/cloudmanager/impl/EnvironmentsTest.java
@@ -32,8 +32,8 @@ import org.apache.commons.io.IOUtils;
 
 import io.adobe.cloudmanager.CloudManagerApiException;
 import io.adobe.cloudmanager.Environment;
+import io.adobe.cloudmanager.EnvironmentLog;
 import io.adobe.cloudmanager.generated.model.LogOptionRepresentation;
-import io.adobe.cloudmanager.model.EnvironmentLog;
 import io.adobe.cloudmanager.model.Variable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
